Tweak console implementation

+++ src/components.rs
@@ -3,12 +3,12 @@ use std::collections::VecDeque;
 use tetra::graphics::Color;
 
 pub struct Position {
-    pub x: i32,
-    pub y: i32,
+    pub x: usize,
+    pub y: usize,
 }
 
 impl Position {
-    pub fn new(x: i32, y: i32) -> Position {
+    pub fn new(x: usize, y: usize) -> Position {
         Position { x, y }
     }
 }
