+++ src/entities.rs
@@ -3,7 +3,7 @@ use tetra::graphics::Color;
 
 use crate::components::*;
 
-pub fn create_player(world: &mut World, x: i32, y: i32) -> Entity {
+pub fn create_player(world: &mut World, x: usize, y: usize) -> Entity {
     world.spawn((
         Position::new(x, y),
         Sprite::new('@', Color::WHITE),
@@ -11,7 +11,7 @@ pub fn create_player(world: &mut World, x: i32, y: i32) -> Entity {
     ))
 }
 
-pub fn create_snake(world: &mut World, x: i32, y: i32) -> Entity {
+pub fn create_snake(world: &mut World, x: usize, y: usize) -> Entity {
     world.spawn((
         Position::new(x, y),
         Sprite::new('S', Color::rgb(0.0, 1.0, 0.0)),
@@ -19,7 +19,7 @@ pub fn create_snake(world: &mut World, x: i32, y: i32) -> Entity {
     ))
 }
 
-pub fn create_rat(world: &mut World, x: i32, y: i32) -> Entity {
+pub fn create_rat(world: &mut World, x: usize, y: usize) -> Entity {
     world.spawn((
         Position::new(x, y),
         Sprite::new('r', Color::rgb(0.59, 0.41, 0.31)),
